v0.01 - hay un mundo procedural de pocas celdas

+++ src/Utils.java
@@ -1,7 +1,7 @@
 package src;
 
 public class Utils {
-    public static int redondear(float n) {
+    public static int redondear(double n) {
         int res = 0;
         while (n >= 1 || n <= -1) {
             n += n > 0 ? -1 : 1;
